+++ src/Proton/Codec/Encoders/AbstractDescribedMapTypeEncoder.cs
@@ -102,7 +102,7 @@ namespace Apache.Qpid.Proton.Codec.Encoders
 
       private void WriteSmallType(IProtonBuffer buffer, IEncoderState state, M value, int elementCount)
       {
-         int startIndex = buffer.WriteOffset;
+         long startIndex = buffer.WriteOffset;
 
          // Reserve space for the size and write the count of list elements.
          buffer.EnsureWritable(sizeof(byte) + sizeof(byte));
@@ -112,14 +112,14 @@ namespace Apache.Qpid.Proton.Codec.Encoders
          WriteMapEntries(buffer, state, value);
 
          // Move back and write the size
-         int writeSize = (buffer.WriteOffset - startIndex) - sizeof(byte);
+         long writeSize = (buffer.WriteOffset - startIndex) - sizeof(byte);
 
          buffer.SetUnsignedByte(startIndex, ((byte)writeSize));
       }
 
       private void WriteLargeType(IProtonBuffer buffer, IEncoderState state, M value, int elementCount)
       {
-         int startIndex = buffer.WriteOffset;
+         long startIndex = buffer.WriteOffset;
 
          // Reserve space for the size and write the count of list elements.
          buffer.EnsureWritable(sizeof(int) + sizeof(int));
@@ -129,9 +129,9 @@ namespace Apache.Qpid.Proton.Codec.Encoders
          WriteMapEntries(buffer, state, value);
 
          // Move back and write the size
-         int writeSize = (buffer.WriteOffset - startIndex) - sizeof(int);
+         long writeSize = (buffer.WriteOffset - startIndex) - sizeof(int);
 
-         buffer.SetInt(startIndex, writeSize);
+         buffer.SetInt(startIndex, (int)writeSize);
       }
 
       public override void WriteArray(IProtonBuffer buffer, IEncoderState state, object[] values)
@@ -140,7 +140,7 @@ namespace Apache.Qpid.Proton.Codec.Encoders
          // Write the Array Type encoding code, we don't optimize here.
          buffer.WriteUnsignedByte(((byte)EncodingCodes.Array32));
 
-         int startIndex = buffer.WriteOffset;
+         long startIndex = buffer.WriteOffset;
 
          // Reserve space for the size and write the count of list elements.
          buffer.WriteInt(0);
@@ -152,14 +152,14 @@ namespace Apache.Qpid.Proton.Codec.Encoders
          WriteRawArray(buffer, state, values);
 
          // Move back and write the size
-         int writeSize = buffer.WriteOffset - startIndex - sizeof(int);
+         long writeSize = buffer.WriteOffset - startIndex - sizeof(int);
 
          if (writeSize > Int32.MaxValue)
          {
             throw new ArgumentOutOfRangeException("Cannot encode given array, encoded size to large: " + writeSize);
          }
 
-         buffer.SetInt(startIndex, writeSize);
+         buffer.SetInt(startIndex, (int)writeSize);
       }
 
       public override void WriteRawArray(IProtonBuffer buffer, IEncoderState state, Object[] values)
@@ -171,7 +171,7 @@ namespace Apache.Qpid.Proton.Codec.Encoders
          {
             M map = (M)values[i];
             int count = GetMapEntries(map);
-            int mapStartIndex = buffer.WriteOffset;
+            long mapStartIndex = buffer.WriteOffset;
 
             // Reserve space for the size and write the count of list elements.
             buffer.WriteInt(0);
@@ -180,9 +180,9 @@ namespace Apache.Qpid.Proton.Codec.Encoders
             WriteMapEntries(buffer, state, map);
 
             // Move back and write the size
-            int writeSize = buffer.WriteOffset - mapStartIndex - sizeof(int);
+            long writeSize = buffer.WriteOffset - mapStartIndex - sizeof(int);
 
-            buffer.SetInt(mapStartIndex, writeSize);
+            buffer.SetInt(mapStartIndex, (int)writeSize);
          }
       }
    }
